requirements.txt
langchain-core
langchain_community 
langgraph 
langchain-groq 
langchain_google_genai
langchain_chroma
langchain_openai
chromadb
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
beautifulsoup4
python-dotenv
colorama
langserve
sse_starlette
uvicorn
gunicorn
fastapi
openai
anthropic
pyyaml